URL Cleanup
This commit updates URLs to prefer the https protocol. Redirects are not followed to avoid accidentally expanding intentionally shortened URLs (i.e. if using a URL shortener).

# Fixed URLs

## Fixed But Review Recommended
These URLs were fixed, but the https status was not OK. However, the https status was the same as the http request or http redirected to an https URL, so they were migrated. Your review is recommended.

* [ ] http://fraud-detection/frauds (UnknownHostException) with 4 occurrences migrated to:
  https://fraud-detection/frauds ([https](https://fraud-detection/frauds) result UnknownHostException).

## Fixed Success
These URLs were switched to an https URL with a 2xx status. While the status was successful, your review is still recommended.

* [ ] http://maven.apache.org/xsd/assembly-1.1.3.xsd with 1 occurrences migrated to:
  https://maven.apache.org/xsd/assembly-1.1.3.xsd ([https](https://maven.apache.org/xsd/assembly-1.1.3.xsd) result 200).

# Ignored
These URLs were intentionally ignored.

* http://localhost:6543/fraud with 1 occurrences
* http://localhost:6544/fraud with 1 occurrences
* http://localhost:6544/message with 1 occurrences
* http://localhost:6545/frauds with 1 occurrences
* http://localhost:6545/message with 1 occurrences
* http://localhost:8083/stubs with 1 occurrences
* http://localhost:8083/triggers with 1 occurrences
* http://localhost:8083/triggers/trigger_a_fraud with 1 occurrences
* http://localhost:8761 with 1 occurrences
* http://localhost:8765 with 1 occurrences
* http://localhost:9876/frauds with 2 occurrences
* http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.3 with 2 occurrences
* http://www.w3.org/2001/XMLSchema-instance with 1 occurrences